Extract shared ListViewController snapshot tests
