+++ EssentialFeed/EssentialFeediOSTests/Share UI/snapshots/EMPTY_LIST_dark.png
