+++ EssentialFeed/EssentialFeediOSTests/Share UI/snapshots/LIST_WITH_ERROR_MESSAGE_light.png
